+++ src/components/Console/Console.tsx
@@ -2,23 +2,31 @@ import { FC, useEffect, useState } from 'react'
 import logo from '../../assets/logo.svg';
 import styled from 'styled-components';
 import { useRef } from 'react';
-import { commandResponse } from '../../data/commands';
+import { WELCOME_MESSAGE, commandResponse } from '../../data/commands';
 import { useDispatch, useSelector } from 'react-redux';
 import CommandLine, { ICommandLine } from './CommandLine';
-import { RootState } from '../../app/store';
-import { runCommandInterpreter } from './commandInterpreter';
+import { CommandInterpreter } from './commandInterpreter';
+import Logo from '../Logo/Logo';
 
 
 const Console : FC = () => {
   const inputRef = useRef<HTMLInputElement>(null);
-  const [commandLog,setCommandLog] = useState<ICommandLine[]>([]);
+  const consoleHistoryRef = useRef<HTMLInputElement>(null);
+  const dispatch = useDispatch();
+  
+  const [commandLog,setCommandLog] = useState<ICommandLine[]>(WELCOME_MESSAGE);
   const [commandHistory,setCommandHistory] = useState<string[]>([]);
   const [previousCommandIndex,setPreviousCommandIndex] = useState<number>(0);
+
+  const clearConsole = () => {
+    setCommandLog([]);
+  }
+
+  const commandInterpreter = new CommandInterpreter(dispatch,clearConsole);
+
   const COMMANDS_LOG_CAPACITY = 70;
   const COMMANDS_HISTORY_CAPACITY = 50;
 
-  const desktopState = useSelector((state: RootState) => state.desktop)
-  const dispatch = useDispatch();
 
   const handleAddCommandLog = (commandLine:ICommandLine) => {
     if(commandLog.length + 1 === COMMANDS_LOG_CAPACITY){
@@ -46,12 +54,8 @@ const Console : FC = () => {
         const command = inputRef.current.value;
         
         //interpret the command and get a response
-        const res : commandResponse | undefined = runCommandInterpreter({
-            input: command,
-            currentDesktopState: desktopState,
-            dispatch: dispatch,
-            clearConsoleAction: ()=>{setCommandLog([])}
-        });
+        const res : commandResponse | undefined = commandInterpreter.run(command);
+
         if(res){
             const cmdLine : ICommandLine = {
                 command: command,
@@ -59,7 +63,7 @@ const Console : FC = () => {
                 infoType: res.code
             }
             handleAddCommandLog(cmdLine);
-            handleAddCommandHistory(cmdLine.command);
+            if(cmdLine.command) handleAddCommandHistory(cmdLine.command);
         }
         //clear the input
         inputRef.current.value="";
@@ -80,6 +84,7 @@ const Console : FC = () => {
 
   useEffect(()=>{
     setPreviousCommandIndex(commandHistory.length);
+    consoleHistoryRef.current?.scrollTo(0,consoleHistoryRef.current.scrollHeight);
   },[commandHistory])
 
   const handleKeyDown = (e:React.KeyboardEvent<HTMLInputElement>) => {
@@ -93,14 +98,17 @@ const Console : FC = () => {
             getPreviousCommand(e.key==="ArrowUp" ? -1 : 1);
             break;
         default:
-            setPreviousCommandIndex(commandLog.length);
+            setPreviousCommandIndex(commandHistory.length);
             break;
     }
   }
 
   return (
     <StyledConsole onClick={focusConsole}>
-        <StyledConsoleHistory>
+        <StyledLogo>
+          <img src={logo} alt="Keyboard Kitchen logo" />
+        </StyledLogo>
+        <StyledConsoleHistory ref={consoleHistoryRef}>
             {
                 commandLog.map((command,index) => (
                     <CommandLine 
@@ -113,9 +121,6 @@ const Console : FC = () => {
             }
         </StyledConsoleHistory>
         <StyledConsoleLabel>
-            <StyledConsoleCommand>
-                What would you like to do?
-            </StyledConsoleCommand>
             <StyledConsoleInput type="text" ref={inputRef} onKeyDown={(e)=>handleKeyDown(e)} />
         </StyledConsoleLabel>
     </StyledConsole>
@@ -134,8 +139,30 @@ export const StyledConsole = styled.div`
 
     color: white;
     background-color: var(--color-console);
-    background: url(${logo}) top 0 right 50% no-repeat;
     background-size: 50%;
+`;
+
+const StyledLogo = styled.div`
+  position: relative;
+
+  & > img {
+    width: 60%;
+    margin: 0 auto;
+  }
+
+  &::before {
+    --fade-element-height: 3rem;
+
+    content: "";
+    position: absolute;
+    z-index: 1;
+    left: 0;
+    bottom: calc(var(--fade-element-height) * -1);
+    
+    background: linear-gradient(180deg, var(--color-console) 0%, rgba(0,0,0,0) 100%);
+    width: 100%;
+    height: var(--fade-element-height);
+  }
 `;
 
 const StyledConsoleHistory = styled.div`
@@ -143,7 +170,7 @@ const StyledConsoleHistory = styled.div`
     flex: 1;
     overflow-y: scroll;
     
-    margin-top: 7rem;
+    /* margin-top: 7rem; */
     padding: 0 1.5rem;
 
   ::-webkit-scrollbar {
@@ -163,27 +190,29 @@ const StyledConsoleHistory = styled.div`
   }
 `;
 
-const StyledConsoleCommand = styled.p`
-    padding: .5rem 0;
-    &::before {
-        content: "> ";
-    }
-`
-
 const StyledConsoleLabel = styled.label`
-    padding: 1rem;
-    border-radius: .5rem;
-    background: #222;
+  padding: 1rem;
+  border-radius: .5rem;
+  background: #222;
+  position: relative;
+
+  &::before {
+    content: ">";
+    position: absolute;
+    left: 1.5rem;
+    top: 50%;
+    transform: translateY(-50%);
+  }
 `
 
 const StyledConsoleInput = styled.input`
     width: 100%;
-    padding-left: 1.5rem;
+    padding-left: 3rem;
     background: none;
     border: none;
     
     color: inherit;
-    font-size: var(--fs-heading);
+    font-size: var(--fs-body);
 
     &:focus {
         outline: none;