+++ src/components/Panels/CookbookPanel/CookbookPanel.tsx
@@ -6,22 +6,22 @@ import { useDispatch } from 'react-redux';
 import { toggleOpenPanel } from '../../../features/desktop/desktopSlice';
 import React from 'react';
 
-const RecipeBookPanel = () => {
+const CookbookPanel = () => {
   const dispatch = useDispatch();
 
   const handlePanelClose = () => {
     dispatch(toggleOpenPanel({
-      panelType: PANELS.RECIPEBOOK,
+      panelType: PANELS.COOKBOOK,
       opened: false
     }))
   }
   
   const recipes = [...RECIPES.keys()].sort((a,b)=>a.localeCompare(b)); // Get all recipes names and sort them alphabetically
   return (
-    <Panel title="Recipe Book" panelType={PANELS.RECIPEBOOK} handlePanelClose={handlePanelClose}>
+    <Panel title="Cookbook" panelType={PANELS.COOKBOOK} handlePanelClose={handlePanelClose}>
         <List type="ol" items={recipes} />
     </Panel>
   )
 }
 
-export default React.memo(RecipeBookPanel)
+export default React.memo(CookbookPanel)